refactor[resource_path]: moved to gui_utils
