+++ config.py
@@ -1,6 +1,6 @@
 """Configuration file. Contains DB_PATH & ICON_PATH"""
 
-from logic.db_logic import resource_path
+from logic.gui_utils import resource_path
 from pathlib import Path
 
 # CHANGE THIS TO USE CUSTOM DB / ИЗМЕНИТЕ ЭТО ДЛЯ СВОЕЙ БД
